+++ shared/data/districts-data.ts
